requirements.txt
python-docx==1.1.2
pdfminer.six==20231228
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
sentence-transformers==2.7.0
openai==1.43.0
python-dotenv==1.0.1
